Removed Warning to require RSkyrimChildren Patch
Given the uncertainty about the necessity of the USKP patch for this
mod, might be best to just leave the warning out till we get some
clarification from the author on this. As states here
(http://forums.bethsoft.com/topic/1503686-rel-loot-thread-10/page-7#entry23710051)
there is no conflict between the USKP and this mod, so the purpose of
the compatibility patch is unclear. Additionally, the patch and the main
file's masters' load order is not in agreement, so there is really no
way for LOOT to properly sort these.