Remove stability warning for Oblivion Hairs.
The mod page linked to no longer exists. A fork exists, but only notes
that running lots of mods will affect stability, so the information is
really too generic to attach to any one mod, and isn't suitable for a
general message, since the limit very much depends on individual
circumstances.