Update for The Eyes of Beauty
While I understand the intent of keeping historical data changes to xEdit 4.0.3 mean that certain ITMs are not being removed. The historical data is causing confusing and need looking into by Elminster.

If Elminster determins that a 4.0.3x needs to contain a fix then additional cleaning information will be added at that time.